requirements.txt
# Core Data Science Libraries
pandas==1.5.3
numpy==1.24.3
scipy==1.10.1

# Machine Learning
scikit-learn==1.3.2
xgboost==1.7.6

# Visualization
matplotlib==3.6.2
seaborn==0.12.2

# Jupyter
jupyterlab==3.5.2
ipykernel

# Network Analysis (if needed for feature engineering)
networkx==3.1
python-igraph==0.10.4
leidenalg==0.9.1

# Text Processing (if needed)
gensim==4.3.2
nltk==3.8.1
wordcloud==1.9.2

# Deep Learning (if needed)
transformers==4.35.2
torch==2.1.0

# Utilities
requests==2.28.1
tqdm==4.66.1

# Excel (used to load proposition metadata from XLSX)
openpyxl==3.1.2

# Figures / slides (used to generate editable pipeline diagram)
python-pptx==0.6.23

